requirements.txt
# ─────────────────────────────────────────────
# SpecAutoAgent - Python 패키지 의존성
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Azure OpenAI SDK
openai==1.51.0

# 환경변수
python-dotenv==1.0.1

# 데이터 검증
pydantic==2.9.2

# 문서 생성
openpyxl==3.1.5       # Excel (.xlsx)
python-docx==1.1.2    # Word (.docx)

# 템플릿 엔진 (Web UI)
jinja2==3.1.4

# 파일 업로드 (UploadFile / multipart form-data 필수)
python-multipart==0.0.12

# HTTP 클라이언트 (openai 내부 의존)
httpx==0.27.2
